boot: use a clean boot_fat.uefi.img
Copy the clean boot_fat.uefi.img. The reason for having a pre-generated
boot_fat.uefi.img here is because on newer versions of mkfs.fat (4.0),
there it seems impossible to set the "Reserved vectors" to "1". It
defaults to 4 reserved sectors and it can go higher but not lower (-R
flag). I could be a bug in the new mkfs.fat tool. So, the pre-build
image here some from the mkfs.fat 3.0.26 (2014-03-07), which gives the
correct number of reserved sectors. This replace the "dd" and the
"mkfs.fat" steps previously done here.

Change-Id: I7bf53cc4d9c619363310387531f24e109e9afc4f
Signed-off-by: Joakim Bech <joakim.bech@linaro.org>